+++ packages/@ember/-internals/views/index.ts
@@ -20,5 +20,4 @@ export { default as ComponentLookup } from './lib/component_lookup';
 export { default as CoreView } from './lib/views/core_view';
 export { default as ActionSupport } from './lib/mixins/action_support';
 export { MUTABLE_CELL } from './lib/compat/attrs';
-export { default as ActionManager } from './lib/system/action_manager';
 export { default as ViewStates } from './lib/views/states';