+++ packages/@ember/reactive/index.ts
@@ -1,8 +1,10 @@
-export {
-  trackedArray,
-  trackedObject,
-  trackedWeakSet,
-  trackedSet,
-  trackedMap,
-  trackedWeakMap,
-} from '@glimmer/validator';
+/**
+ * The `@ember/reactive` package contains common reactive utilities
+ * for tracking values and creating reactive data structures.
+ *
+ * @module @ember/reactive
+ * @public
+ */
+
+// empty module (for now)
+export default {};